se crea el metodo para crear token

+++ models/message.go
@@ -2,7 +2,7 @@ package models
 
 //Message estructura de los mensajes de respuesta
 type Message struct {
-	Code    uint16 `json:"code"`
+	Code    int    `json:"code"`
 	Data    string `json:"data,omitempty"`
 	Message string `json:"message,omitempty"`
 }
